requirements.txt
django
django-environ
psycopg2==2.9.5
django-import-export==3.0.1
pandas